Fix CategoryLabel Html issue (#42)
